renamed&fixed deleteModal to teamDropList

+++ src/components/Team/TeamTable.vue
@@ -1,58 +1,109 @@
 <template>
-  <div class="fe-team__table">
-    <div class="fe-table__container">
-      <div class="spinner-icon" v-if="loading">
-        <spinner />
-      </div>
-      <v-data-table
-        :headers="headers"
-        :items="tableItems"
-        :options.sync="options"
-        :loading="loading"
-        :items-per-page="pageSize"
-        hide-default-footer
-      >
-        <template v-slot:item.profileImageUrl="{ item }">
-          <div class="d-flex justify-center">
-            <img :src="item.profileImageUrl" alt="icon" class="fe-table-cell__avatar-img" />
-          </div>
-        </template>
-        <template v-slot:item.roleIds="{ item }">
-          <div class="fe-flex fe-full-width fe-flex-no-wrap" v-if="item.roleIds && item.roleIds.length > 0">
-            <div class="fe-flex">
-              <div
-                class="fe-tag fe-mr-1 fe-mb-1 fe-mt-1 fe-tag-default fe-tag-small"
-                v-for="role in item.roleIds"
-                :key="role"
-              >
-                {{ getRoleName(role) }}
+  <div>
+    <div class="fe-team__table">
+      <div class="fe-table__container">
+        <div class="spinner-icon" v-if="loading">
+          <spinner />
+        </div>
+        <v-data-table
+          :headers="headers"
+          :items="tableItems"
+          :options.sync="options"
+          :loading="loading"
+          :items-per-page="pageSize"
+          :fixed-header="true"
+          hide-default-footer
+        >
+          <template v-slot:item.profileImageUrl="{ item }">
+            <div class="d-flex justify-center">
+              <img
+                :src="item.profileImageUrl"
+                alt="icon"
+                class="fe-table-cell__avatar-img"
+              />
+            </div>
+          </template>
+          <template v-slot:item.roleIds="{ item }">
+            <div
+              class="fe-flex fe-full-width fe-flex-no-wrap"
+              v-if="item.roleIds && item.roleIds.length > 0"
+            >
+              <div class="fe-flex">
+                <div
+                  class="fe-tag fe-mr-1 fe-mb-1 fe-mt-1 fe-tag-default fe-tag-small"
+                  v-for="role in item.roleIds"
+                  :key="role"
+                >
+                  {{ getRoleName(role) }}
+                </div>
               </div>
             </div>
-          </div>
-        </template>
-        <template v-slot:item.createdAt="{ item }">
-          <div class="datetime" v-if="item.createdAt">
-            <span class="date">{{ parseDate(item.createdAt) }}</span>
-            <span class="left-time">{{ leftTime(item.createdAt) }} days ago</span>
-          </div>
-          <div class="fe-tag fe-tag-primary fe-tag-small" v-else>
-            <span class="date">{{ $t("common.pendingApproval") }}</span>
-          </div>
-        </template>
-        <template v-slot:item.lastLogin="{ item }">
-          <div class="datetime" v-if="item.lastLogin">
-            <span class="date">{{ parseDate(item.lastLogin) }}</span>
-            <span class="left-time"
-              >{{ leftTime(item.lastLogin) ? leftTime(item.lastLogin) + "days ago" : "today" }}
-            </span>
-          </div>
-          <span v-else>N/A</span>
-        </template>
-        <template v-slot:item.id="{ item }">
-          <DeleteModal @deleteUser="deleteUser(item.id)"></DeleteModal>
-        </template>
-      </v-data-table>
+          </template>
+          <template v-slot:item.createdAt="{ item }">
+            <div class="datetime" v-if="item.createdAt">
+              <span class="date">{{ parseDate(item.createdAt) }}</span>
+              <span class="left-time"
+                >{{ leftTime(item.createdAt) }} days ago</span
+              >
+            </div>
+            <div class="fe-tag fe-tag-primary fe-tag-small" v-else>
+              <span class="date">{{ $t("common.pendingApproval") }}</span>
+            </div>
+          </template>
+          <template v-slot:item.lastLogin="{ item }">
+            <div class="datetime" v-if="item.lastLogin">
+              <span class="date">{{ parseDate(item.lastLogin) }}</span>
+              <span class="left-time"
+                >{{
+                  leftTime(item.lastLogin)
+                    ? leftTime(item.lastLogin) + "days ago"
+                    : "today"
+                }}
+              </span>
+            </div>
+            <span v-else>N/A</span>
+          </template>
+          <template v-slot:item.id="{ item }">
+            <TeamDropList
+              @deleteUser="userDeleteModal(item.id)"
+              :sendEmail="!item.lastLogin"
+            ></TeamDropList>
+          </template>
+        </v-data-table>
+      </div>
     </div>
+    <FModal
+      :open-modal="showUserDeleteModal"
+      :onCloseModal="onCloseModal"
+      :headText="'Delete team member'"
+    >
+      <template #content>
+        <div class="fe-dialog-body">
+          <span
+            >You are about to remove {{ textUserDeleteModal }} from this
+            account. Are you sure?</span
+          >
+        </div>
+
+        <v-card-actions>
+          <v-btn
+            :class="{ 'fe-button-disabled': loading }"
+            text
+            @click="onCloseModal"
+            class="fe-button"
+          >
+            {{ $t("common.cancel") }}
+          </v-btn>
+          <v-btn
+            text
+            class="fe-button fe-button-large fe-button-danger"
+            @click="deleteUser"
+          >
+            {{ $t("common.delete") }}
+          </v-btn>
+        </v-card-actions>
+      </template>
+    </FModal>
   </div>
 </template>
 
@@ -61,68 +112,70 @@ import Vue from "vue";
 import { mapState } from "@/plugins/fronteggCore/map-state";
 import { AuthState } from "@/plugins/fronteggAuth/Api";
 import Spinner from "@/components/Common/Spinner.vue";
-import DeleteModal from "@/components/core/elements/Modal/DeleteModal.vue";
+import TeamDropList from "./TeamDropList.vue";
+import FModal from "@/components/core/elements/Modal/FModal.vue";
 
 import { teamActions } from "@/plugins/fronteggAuth/Api/TeamState/index.ts";
 
 import { FRONTEGG_STORE_KEY } from "@/plugins/fronteggCore/constants";
 
-import { TableOptions } from './interfaces';
-
+import { TableOptions } from "./interfaces";
 
 export default Vue.extend({
   name: "TeamTable",
-  components: { Spinner, DeleteModal },
+  components: { Spinner, TeamDropList, FModal },
   props: {
     page: {
       default: 0,
-      type: Number
-    }
+      type: Number,
+    },
   },
   data() {
     return {
       ...mapState(this, {
-        teamState: (state: { auth: AuthState }) => state.auth.teamState
+        teamState: (state: { auth: AuthState }) => state.auth.teamState,
       }),
+      showUserDeleteModal: false,
+      textUserDeleteModal: "",
+      idUserDeleteModal: null,
       options: {} as TableOptions,
-
       headers: [
         {
           text: "",
           sortable: false,
-          value: "profileImageUrl"
+          value: "profileImageUrl",
         },
         {
           text: "Name",
           sortable: true,
-          value: "name"
+          value: "name",
         },
         {
           text: "Email",
           sortable: true,
-          value: "email"
+          value: "email",
         },
         {
           text: "Roles",
           sortable: false,
-          value: "roleIds"
+          value: "roleIds",
         },
         {
           text: "Joined Team",
           sortable: true,
-          value: "createdAt"
+          value: "createdAt",
         },
         {
           text: "Last Login",
           sortable: true,
-          value: "lastLogin"
+          value: "lastLogin",
         },
         {
           text: "",
           sortable: false,
-          value: "id"
-        }
-      ]
+          value: "id",
+        },
+      ],
     };
   },
   computed: {
@@ -137,45 +190,61 @@ export default Vue.extend({
     },
     loading() {
       return this.teamState.loaders.USERS;
-    }
+    },
   },
   watch: {
     options: {
       handler() {
         this.$emit("fetchTableData", this.options);
       },
-      deep: true
+      deep: true,
     },
     page(val) {
       this.options.page = val;
-    }
+    },
   },
   methods: {
+    userDeleteModal(id) {
+      this.showUserDeleteModal = true;
+      this.idUserDeleteModal = id;
+      this.textUserDeleteModal = this.tableItems.find(
+        (item) => item.id === id
+      ).email;
+    },
+    onCloseModal() {
+      this.showUserDeleteModal = false;
+    },
     getRoleName(roleId: string) {
-      const roleObj = this.roles.find(i => i.id === roleId);
+      const roleObj = this.roles.find((i) => i.id === roleId);
       return roleObj ? roleObj.name : null;
     },
     parseDate(date: string) {
-      return this.$moment(date).format('LLLL');
+      return this.$moment(date).format("LLLL");
     },
     leftTime(date: string) {
-      return new Date(new Date().getTime() - new Date(date).getTime()).getUTCDate() - 1;
-    },
-    deleteUser(id: string) {
-      this?.[FRONTEGG_STORE_KEY]?.dispatch(
-        teamActions.deleteUser({
-          userId: id
-        })
+      return (
+        new Date(new Date().getTime() - new Date(date).getTime()).getUTCDate() -
+        1
       );
+    },
+    deleteUser() {
+      if (this.idUserDeleteModal) {
+        this?.[FRONTEGG_STORE_KEY]?.dispatch(
+          teamActions.deleteUser({
+            userId: this.idUserDeleteModal,
+          })
+        );
+        this.onCloseModal();
+      }
     },
     resendActivationLink(id: string) {
       this?.[FRONTEGG_STORE_KEY]?.dispatch(
         teamActions.resendActivationLink({
-          userId: id
+          userId: id,
         })
       );
-    }
-  }
+    },
+  },
 });
 </script>
 
@@ -216,7 +285,13 @@ export default Vue.extend({
   border-bottom: 3px solid transparent;
   transition: border 0.3s;
 }
-.fe-table__container .v-data-table .v-data-table__wrapper table thead.v-data-table-header tr th {
+.fe-table__container
+  .v-data-table
+  .v-data-table__wrapper
+  table
+  thead.v-data-table-header
+  tr
+  th {
   display: table-cell;
   background: var(--color-gray-0);
   border-bottom: none;
